refactor: egoist.generate.go
